+++ lib/lockbox/utils.rb
@@ -14,6 +14,15 @@ module Lockbox
         options[:key] = Lockbox.attribute_key(table: table, attribute: attribute, master_key: options.delete(:master_key))
       end
 
+      if options[:previous_versions].is_a?(Array)
+        options[:previous_versions] = options[:previous_versions].dup
+        options[:previous_versions].each_with_index do |version, i|
+          if !(version[:key] || version[:encryption_key] || version[:decryption_key]) && version[:master_key]
+            options[:previous_versions][i] = version.merge(key: Lockbox.attribute_key(table: table, attribute: attribute, master_key: version.delete(:master_key)))
+          end
+        end
+      end
+
       Lockbox.new(**options)
     end