+++ test/support/active_record.rb
@@ -27,7 +27,7 @@ class User < ActiveRecord::Base
 
   mount_uploader :document, DocumentUploader
 
-  encrypts :email, previous_versions: [{key: Lockbox.generate_key}]
+  encrypts :email, previous_versions: [{key: Lockbox.generate_key}, {master_key: Lockbox.generate_key}]
 
   key_pair = Lockbox.generate_key_pair
   encrypts :phone, algorithm: "hybrid", encryption_key: key_pair[:encryption_key], decryption_key: key_pair[:decryption_key]