+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.2 (unreleased)
 
 - Added `encode` option to `Lockbox::Encryptor`
+- Added support for `master_key` in `previous_versions`
 - Added `Lockbox::Audit` model
 
 ## 0.3.1 (2019-12-26)