+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Added `encode` option to `Lockbox::Encryptor`
 - Added support for `master_key` in `previous_versions`
-- Added `Lockbox::Audit` model
+- Added experimental `Lockbox::Audit` model
 
 ## 0.3.1 (2019-12-26)